refactor(test): split preview document tests by format family
