+++ src/preview/tests/documents/mod.rs
@@ -1,0 +1,5 @@
+use super::*;
+
+mod epub;
+mod office;
+mod pdf;
